requirements.txt.txt
streamlit
scikit-learn
pandas
numpy
joblib
Pillow
